+++ package.json
@@ -9,11 +9,11 @@
   "bugs": {
     "url": "https://github.com/menmos/menmos-web/issues"
   },
+  "source": "src/pages/index.html",
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "export": "next export",
-    "start": "next start",
+    "start": "parcel",
+    "build": "parcel build",
+    "serve": "parcel serve",
     "test": "jest",
     "typecheck": "tsc --noEmit",
     "lint": "eslint . && stylelint '**/*.scss'",
@@ -27,18 +27,21 @@
     "@fortawesome/react-fontawesome": "^0.1.16",
     "axios": "^0.21.4",
     "bricks.js": "^1.8.0",
+    "history": "^5.2.0",
     "mime": "^3.0.0",
     "moment": "^2.29.1",
-    "next": "11.1.2",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "react-router-dom": "^6.2.1",
     "sass": "^1.39.2"
   },
   "devDependencies": {
-    "@next/eslint-plugin-next": "^11.1.2",
+    "@parcel/transformer-sass": "^2.2.1",
+    "@parcel/transformer-typescript-tsc": "^2.2.1",
     "@types/bricks.js": "^1.8.2",
     "@types/mime": "^2.0.3",
     "@types/react": "^17.0.20",
+    "@types/react-dom": "^17.0.11",
     "@typescript-eslint/eslint-plugin": "^4.31.0",
     "@typescript-eslint/parser": "^4.31.0",
     "eslint": "^7.32.0",
@@ -49,6 +52,7 @@
     "eslint-plugin-react-hooks": "^4.2.0",
     "eslint-plugin-unicorn": "^36.0.0",
     "jest": "^27.1.1",
+    "parcel": "2.2.1",
     "prettier": "^2.4.0",
     "stylelint": "^13.13.1",
     "stylelint-config-prettier": "^8.0.2",
@@ -58,5 +62,6 @@
   },
   "engines": {
     "node": ">=16"
-  }
+  },
+  "browserslist": "> 0.5%, last 2 versions, not dead"
 }
