--- pages/_app.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import { AppProps } from "next/app";
-import "../styles/globals.scss";
-import { config } from "@fortawesome/fontawesome-svg-core";
-import "@fortawesome/fontawesome-svg-core/styles.css";
-config.autoAddCss = false;
-
-function MyApp({ Component, pageProps }: AppProps): JSX.Element {
-  return <Component {...pageProps} />;
-}
-
-export default MyApp;
