+++ src/components/content.tsx
@@ -1,7 +1,7 @@
 import React, { FC, useCallback, useEffect, useState } from "react";
 
-import { Grid } from "../components/grid";
-import { query, DEFAULT_PARAMS, Blob } from "../src/api/query";
+import { Grid } from "./grid";
+import { query, DEFAULT_PARAMS, Blob } from "../api/query";
 import { Card } from "./card/card";
 
 interface Properties {
