+++ src/styles/blobs.module.scss
