+++ src/styles/footer.module.scss
