+++ src/styles/globals.scss
@@ -1,6 +1,6 @@
 html,
 body,
-#__next {
+#app {
   padding: 0;
   margin: 0;
   height: 100%;
