+++ src/styles/spinner.module.scss
