+++ src/utils/history.ts
@@ -1,0 +1,2 @@
+import { createBrowserHistory } from "history";
+export default createBrowserHistory();
