+++ tsconfig.json
@@ -21,9 +21,10 @@
     "isolatedModules": true,
     "jsx": "preserve",
     "allowJs": false,
-    "resolveJsonModule": true
+    "resolveJsonModule": true,
+    "experimentalDecorators": true
   },
 
-  "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx"],
+  "include": ["**/*.ts", "**/*.tsx"],
   "exclude": ["node_modules"]
 }
